voice-command/requirements.txt
SpeechRecognition
PyYAML
pyaudio
pocketsphinx